+++ linqex/__init__.py
@@ -1,4 +1,4 @@
-from linqex.build import *
+from linqex.base import *
 from linqex.linq import *
 
-version = "1.6"
+version = "1.6.1"