+++ test.py
@@ -1,0 +1,23 @@
+import sys
+sys.dont_write_bytecode = True
+
+import tests.dict.edit_test
+import tests.dict.filter_test
+import tests.dict.get_test
+import tests.dict.math_test
+import tests.dict.test_data
+import tests.dict.tools_test
+
+import tests.list.edit_test
+import tests.list.filter_test
+import tests.list.get_test
+import tests.list.math_test
+import tests.list.test_data
+import tests.list.tools_test
+
+import tests.item.edit_test
+import tests.item.filter_test
+import tests.item.get_test
+import tests.item.math_test
+import tests.item.test_data
+import tests.item.tools_test